角色，权限新增接口

+++ rain-user/src/main/java/com/rao/dao/system/RainRoleDao.java
@@ -36,6 +36,6 @@ public interface RainRoleDao {
      * @param rainRole
      * @return
      */
-    Integer insertSelective(RainRole rainRole);
+    Long insertSelective(RainRole rainRole);
 
 }
